[gestaltd] split daemon e2e tests into their own package
The daemon coverage run was the long pole in CI/CD: a single serial
`go test -coverprofile ./...` where the daemon package dominated. ~105s of
that came from tests that exec the built gestaltd binary (TestE2E* and the
provider package/release/publish flows) — slow and, because the code runs
out-of-process, contributing ~0 coverage.

Move every binary-exec'ing test into a new internal/daemon/e2e package
(black-box, package e2e) with its own TestMain that builds the gestaltd
binary + provider fixtures. The in-process unit tests (lock/sync, release
metadata/archive helpers, provider-local session) stay in package daemon
with a slim TestMain that builds only the provider fixtures they need.
daemon_internals_test.go locally mirrors the few unexported types needed to
decode the binary's JSON output, keeping the suite black-box. A focused
TestRunProviderPublishCommand unit test keeps the publish command-runner
covered in-process. The bigquery e2e test now resolves the plugin tree via
testutil.RepoRootPath instead of a fixed ../../.. that broke with the deeper
package path (it had started always skipping).

Result: daemon unit + coverage runs in ~7s (was ~44s) at the same 26.4%
statement coverage; the e2e suite (78 tests) runs separately.

CI/CD wiring:
- validate-gestaltd.yml: the single Go Tests job runs unit tests with coverage
  (the separate coverage job is folded in) — affected packages on PRs (still
  compile-checking everything via go vet), all packages on schedule/main. It
  excludes internal/daemon/e2e. New run-e2e input + dedicated e2e job; race
  shards exclude e2e.
- ci.yml: PRs run the Go Tests (unit + coverage) job; race + e2e run on the
  hourly health check.
- cd.yml: run-e2e: true, so e2e is part of validation and gates the CI image
  publish (build-and-push already gates on needs.validate.result).
