+++ gestaltd/internal/daemon/e2e/agent_local_test.go
@@ -1,4 +1,4 @@
-package daemon
+package e2e
 
 import (
 	"encoding/json"
