+++ gestaltd/internal/daemon/e2e/provider_lifecycle_test.go
@@ -1,4 +1,4 @@
-package daemon
+package e2e
 
 import (
 	"crypto/sha256"
